Apply the logger config by label
In order to support labels from Juju, we would like to be able to
configure loggers by their labels. So if a logger has a label of a
model namespace (UUID) then searching from the root module to set the
level of a module should be possible.

As you can configure both module names and labels at the same time,
this becomes tricky. We don't want users to be able to configure a
logger via a label directly and the label should always take
precedence over a user input.